Revert "Merge pull request #373 from au-research/feature/RAID-461"
This reverts commit 1b5777e6d8d2a184513f373867c47b2996861090, reversing
changes made to ad0674f53b1dad6590904be3bdb7cc08492e474a.

+++ api-svc/raid-api/src/main/java/au/org/raid/api/service/OrcidIntegrationService.java
@@ -63,7 +63,7 @@ public class OrcidIntegrationService {
     }
 
     public void updateOrcidRecord(final RaidDto raid) {
-        final var servicePointId = raid.getIdentifier().getOwner().getServicePoint().longValue();
+        final var servicePointId = raid.getIdentifier().getOwner().getServicePoint();
 
         final var servicePoint = servicePointService.findById(servicePointId)
                 .orElseThrow(() -> new ServicePointNotFoundException(servicePointId));
